Fix two glitches in the S2F switch module.

+++ main/S2fSwitch.h
@@ -11,6 +11,7 @@
 #include "protocol/ClockIntf.h"
 
 #include <driver/gpio.h>
+#include <cinttypes>
 
 class S2fSwitch final : public Clock_I
 {
@@ -29,15 +30,15 @@ private:
     // active auto switch filter
     bool (*auto_plug)();
     bool _auto_state = true;
-    int _auto_lag;
-    int _lag_counter = 0;
+    int16_t _auto_lag;
+    int16_t _lag_counter = 0;
 
     // switch polling
-    int _active_level = 0;
+    int16_t _active_level = 0;
     bool _state = false;
     bool _lastButtonRead = false;
-    int _debounce = 0;
-    int _dcount = 4;
+    int16_t _debounce = 0;
+    int16_t _dcount = 4;
 };
 
 extern S2fSwitch *S2FSWITCH;
